+++ lib/easytest/utils.rb
@@ -10,5 +10,15 @@ module Easytest
       # https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda
       "\e]8;;file://#{path}\e\\#{Rainbow(dir).dimgray}#{base}\e]8;;\e\\"
     end
+
+    def raise_if_no_test_name(name, method:)
+      if name.nil? || name.empty?
+        raise FatalError.new("`#{method}` requires a name")
+      end
+    end
+
+    def pluralize(singular, count)
+      count == 1 ? singular : "#{singular}s"
+    end
   end
 end